Create self_logs.go
